requirements.txt
fastapi==0.100.1
uvicorn[standard]==0.23.2
requests
beautifulsoup4
pillow
pydantic
openpyxl
